First step in reoganizing class structure.
Updated the build file so that it prints out status of each test.

Added a directory /site/textFiles which contains text files.

Created a FileIO for getting the contents of text files.

Created a HttpHandlerSelector class. Which is not finished, it needs to
be broken up into separate  classes.

Created an HttpReaderWriter class, that gets HttpRequests from an IO and
send HttpResponses to that same IO.

Created a basic HttpResponse class. It will have to be modified in the
future, as for now it just stores a string.

Updated the HttpServer and HttpServerRunner to use the HttpReaderWriter
and HttpHandlerSelector classes.