Added InvalidHttpRequestException.
Created a custom exceptoin for the HttpRequestBuilder to throw when it
can not create a valid HttpRequest. Moved the HttpVerifier to the
HttpRequestBuilder.

Before the HttpReaderWriter class had to ask the HttpVerifier if the
message it recieved was a valid http request before it could tell the
builder to create a HttpRequest object.

Now it just tells the builder to create a HttpRequest object, if the
builder can not do that it throws an exception.