Removed catch IndexOutOfBoundsException.
The code now checks to see if the array is the correct size before
trying to access an index.

Broke the buildHeaderMap method up, by adding two helper functions.

One to parse the rawRequest into a String array that contains each line
in the header.

And antoher method that takes a String array and check each line to make
sure it is in the formate key: value then addes that key value pair to
the map. If it is not it thorws an InvalidHttpRequestException.