Added ability to server images / pdf files.
also finished reorganization as well as:

Added the ability to pass in a path to be used as the root directory of
the server.

Changed FileIO class so that it no longer returns a String containing
the contents of a file, it now returns a byte array of the file
contents.

Added the ability to server images files (.jpg and .gif) and pdf files.

HttpReaderWriter now takes a HttpResponse as a parameter for it’s send
sendHttpResponse method.

HttpResponse class how stores a byte[] body and a HashMap<String,
String> containing the headers.

Removed the storing of headers in the HttpResquest class as they were
not being used.

Removed the textFiles dir from ./site and put all content to be severed
into ./site