Made changes based on discussion with Darek
Changed contentType method in HttpHandleRequest method to use a HashMap
instead of a bunch of if else blocks.

Changed try block to only contain the code that could throw exception.

Changed generateResponse method in HttpHandlerSelector to return a valid
request stating that the current method isn’t supported yet for non get
methods. Added a comment to remove this once the support for those other
methods is added, as this will be come dead code.