Added Feature to show dir contents with links.
Added a method to FileIO that returns a String[] of the contents of a
directory.

Created class DirListHtml class to build and Html page out of a String[]
of files.

Created handler HandlerGetDirectoryContents that uses DirListHtml to
generate response body.

Modified HandlerRouter so that it returns the
HandlerGetDirectoryContents handler when HttpRequest is for a directory.

Added a sub directory with a text file in it to the ./site folder.