Feature POST resources
Created HanlderPostFileContents class.

Added that handler to the HanlderRouter.

Updated IMessageIO interface and SocketIO class with method to get a
byte[] stream

Modified HttpRequestBuilder so that instead of being supplied a String
containing a rawRequest, it gets passed a reference to the SocketIO
class. This way after parsing the request line and headers it can check
to see if it needs to get a body.

HttpReaderWriter updated to work with the changes made to
HttpRequestBuilder.

HttpRequest now stores a byte[] body, HashMap of headers, and has a
method to get the ContentLength.

Updated MockFileIO for use in tests.