setBody method now requires a content type.
This is to ensure that a Content-Type header gets set if there is a
body. Before the handler building a response would be responsible for
making a seperate method call to setHeader. This change removes that
extra responsiblity and makes setting a Content Typer mandatory if a
body is set.