Create a ICallable interface
Created an ICallable interface with a single method call() that returns
an Object[3] with an int response code, HashMap<String, String> headers,
and byte[] body. This is in preparation for the .jar

This is so that someone using the jar only has to implement a single
method, call(), for their response object; instead of having to create a
response object that has a getResponseCode, getHeaders, getBody.