Add the ability to create custom routes
This now ads the ability for custom routes to be created.
-
Added a CustomRoute class.
Moved the logic in HandlerRouter into a DefaultHandlerRouter class.
Created a CustomHandlerRouter class which CustomRoutes can be added to.
When the CustomHandlerRouter is asked to get a IHandler it will look in
the CustomHandlers first and if it cannot find one it will then ask the
DefaultHandlerRouter to get a IHandler.