Chagned api so that user can create own router
Remvoed the CustomRouter and Route class.

Added a main Router class that takes in the default IHandlerRouter
class in constructor. Then has a method to add a custom IHandlerRouter.

Updated various classes to use Router instead of IHandlerRouter.

These changes were done so that when the project is extracted into a jar
file the user can create there own custom router.