Changed SocketInput to use a scanner.
The prior implementation wasn’t working for some reason. It would hang
when trying to read a request body. I believe it has something to do
with the way I was creating a new input stream in each of my reader
methods.

I changed it so that instead the constructor creates an input stream and
a scanner and both the reader methods use that same scanner.