Added the ability to use wildcard for path
Custom routes can now use a wildcard for the path. This was done so that
users of the server API have the ability to make a custom route that
overrides a method regardless of the path.

This will be required for passing some of the cobspec tests in the
clojure application. (ie HEAD, OPTIONS, etc)